MyPage 리팩토링 및 이미지 경로 수정
