+++ src/pages/MyPage.tsx
@@ -1,26 +1,23 @@
 import { Title } from "../components/common/Title";
-import { TitleAndDesc } from "../components/common/TitleAndDesc";
 import { BackgroundCover } from "../components/common/BackgroundCover";
 import { Modal } from "../components/common/Modal";
 import { useState } from "react";
-import { cacheChargeModalStateFn } from "../Business/Mypage/MyPageModalState";
 import { useEffect, useRef } from "react";
-import { changeProfile, removeProfile } from "../Business/Mypage/MyPageChangeProfile";
-import { setEditClass, mypageBlueEventHandler } from "../Business/Mypage/MypageFouseHandler";
 import { Link } from "react-router-dom";
 import { QueryObserverResult, RefetchOptions, RefetchQueryFilters, useQuery } from "react-query";
 import { fetchAgentMyPageData } from "../fetch/get/agent";
 import { fetchCustomerMyPageData } from "../fetch/get/customer";
-import { fetchRemoveUserProfile } from "../fetch/delete/customer";
-import { fetchChangeName } from "../fetch/put/customer";
-import { fetchAgentEditName } from "../fetch/put/agent";
+import { MyPageProfileImgCompo } from "./MyPageProfileImgCompo";
+import { MyPageProfileButtonCompo } from "./MyPageProfileButtonCompo";
+import { MyPageProfileInfoCompo } from "./MyPageProfileInfoCompo";
+import { MyPageChargeList } from "./customer/MyPageChargeList";
 import styled from "@emotion/styled";
 
 export type reFetchingType = <TPageData>(
   options?: (RefetchOptions & RefetchQueryFilters<TPageData>) | undefined,
 ) => Promise<QueryObserverResult<any, unknown>>;
 
-type fetchDataType = {
+export type fetchDataType = {
   status: string;
   data: {
     id: number;
@@ -45,7 +42,6 @@ export const MyPage = () => {
   const cacheChargeButtonDom = useRef<HTMLButtonElement>(null);
   const pullCacheButtonDom = useRef<HTMLButtonElement>(null);
   const imageDom = useRef<HTMLImageElement>(null);
-  const nameInputDom = useRef<HTMLInputElement>(null);
   const userType = localStorage.getItem("userType");
   const { data, refetch } = useQuery(
     "getData",
@@ -98,13 +94,13 @@ export const MyPage = () => {
         <div className="mb-12 relative">
           <Title>마이페이지</Title>
           {OwnerState && (
-            <ShowMyOfficeButotn>
+            <ShowMyOfficeButton>
               <button className="w-52 btn btn-outline btn-primary">
                 <Link to={"/MyOffice"} className="w-52 btn btn-outline btn-primary absolute right-0 top-0">
                   나의 지점 보기
                 </Link>
               </button>
-            </ShowMyOfficeButotn>
+            </ShowMyOfficeButton>
           )}
         </div>
         <div
@@ -113,144 +109,36 @@ export const MyPage = () => {
           } mb-8 px-8 pb-8 sm:flex-col lg:flex-row lg:gap-x-9`}
         >
           <div className=" sm:w-full lg:w-full">
-            <figure className="w-full h-96 overflow-hidden rounded-lg mb-4">
-              <img
-                src={
-                  fetchUserData && fetchUserData.data.profileImagePath !== "None"
-                    ? fetchUserData.data.profileImagePath
-                    : "public/officeImg/noimage.png"
-                }
-                alt="프로필 이미지"
-                className="w-full h-full object-cover"
-                ref={imageDom}
-              />
-            </figure>
-            <div className="flex justify-between relative sm:px-4 md:px-14 lg:px-10 xl:px-14">
-              <input
-                type="file"
-                id="EditProfile"
-                className="hidden"
-                accept="image/jpeg, image/png"
-                name="profile"
-                readOnly
-                onChange={event => changeProfile(event, imageDom, OwnerState)}
-              />
-              <ProfilePseudoElements htmlFor="EditProfile" className="cursor-pointer">
-                프로필 등록
-              </ProfilePseudoElements>
-              <button
-                onClick={() => {
-                  removeProfile(imageDom);
-                  fetchRemoveUserProfile();
-                }}
-              >
-                프로필 제거
-              </button>
-            </div>
+            <MyPageProfileImgCompo fetchUserData={fetchUserData} imageDom={imageDom} />
+            <MyPageProfileButtonCompo imageDom={imageDom} OwnerState={OwnerState} />
           </div>
-          <div className="w-6/12 sm:w-full lg:w-full">
-            <div className="mb-6 pb-4 border-b border-solid border-accent">
-              <TitleAndDesc
-                title="아이디"
-                description={`${fetchUserData ? fetchUserData.data.email : "불러오기에 실패하였습니다"}`}
-              />
-            </div>
-            <div className="mb-6 pb-4 border-b border-solid border-accent relative leading-loose">
-              <h3 className="font-black text-lg pl-4">닉네임</h3>
-              <form
-                action="#"
-                method="post"
-                onSubmit={event =>
-                  OwnerState === true ? fetchAgentEditName(event, refetch) : fetchChangeName(event, refetch)
-                }
-                encType="multipart/form-data"
-              >
-                {/* defaultValue값에 false일때 값을 주게 되면 해당 값으로 고정 되어버려 넣지 않음 */}
-                {fetchUserData ? (
-                  <input
-                    type="text"
-                    id="EditIdInput"
-                    className="text-base pl-4 mr font-bold bg-transparent"
-                    defaultValue={`${fetchUserData.data.name}`}
-                    readOnly
-                    ref={nameInputDom}
-                    onBlur={() => {
-                      mypageBlueEventHandler(nameInputDom);
-                    }}
-                  />
-                ) : (
-                  <span className="text-base pl-4 mr font-bold bg-transparent">불러오기에 실패했습니다</span>
-                )}
-              </form>
-
-              <EditPosition
-                className="border-b border-solid cursor-pointer"
-                htmlFor="EditIdInput"
-                onClick={() => {
-                  setEditClass(nameInputDom);
-                }}
-              >
-                수정
-              </EditPosition>
-            </div>
-            <div className="mb-6 pb-4 border-b border-solid border-accent">
-              <TitleAndDesc
-                title="현재 포인트"
-                description={`${fetchUserData ? fetchUserData.data.point : "불러오기에 실패하였습니다"}`}
-                type="point"
-              />
-            </div>
-            {fetchUserData?.data.histories && (
-              <div className="flex justify-between gap-x-9">
-                <button
-                  ref={cacheChargeButtonDom}
-                  className={`btn btn-primary grow charge ${cacheChargeModalState && "active"}`}
-                  onClick={() => {
-                    cacheChargeModalStateFn(
-                      cacheChargeModalState,
-                      pullCacheModalState,
-                      setcacheChargeModalState,
-                      setpullCacheModalState,
-                    );
-                  }}
-                >
-                  충전
-                </button>
-              </div>
-            )}
-          </div>
+          <MyPageProfileInfoCompo
+            fetchUserData={fetchUserData}
+            refetch={refetch}
+            OwnerState={OwnerState}
+            cacheChargeModalState={cacheChargeModalState}
+            pullCacheModalState={pullCacheModalState}
+            setcacheChargeModalState={setcacheChargeModalState}
+            setpullCacheModalState={setpullCacheModalState}
+          />
         </div>
-        {fetchUserData?.data.histories && (
-          <div className="pr-10 pl-20 font-bold relative ">
-            <div className="flex justify-end mt-5 sm:justify-center sm:w-full lg:justify-end">
-              <ChargeList>
-                <h4 className="mb-4">충전 내역</h4>
-                <table className=" text-left border-separate border-spacing-y-1.5 sm:whitespace-nowrap sm:border-spacing-y-4 sm:w-full">
-                  <thead className="font-black text-lg">
-                    <tr>
-                      <th className="pr-40 sm:pr-0">날짜</th>
-                      <th>충전 금액</th>
-                    </tr>
-                  </thead>
-                  <tbody className="font-thin text-base">
-                    {fetchUserData?.data.histories &&
-                      fetchUserData.data.histories.map((item, index) => {
-                        const formatDate = item.createdAt.slice(0, 10);
-                        return (
-                          <tr key={index}>
-                            <td>{formatDate}</td>
-                            <td>
-                              {item.chargeAmount}
-                              <span>원</span>
-                            </td>
-                          </tr>
-                        );
-                      })}
-                  </tbody>
-                </table>
-              </ChargeList>
-            </div>
-          </div>
+        {fetchUserData?.data.histories ? (
+          <MyPageChargeList fetchUserData={fetchUserData} />
+        ) : (
+          <table className=" text-left border-separate border-spacing-y-1.5 sm:whitespace-nowrap sm:border-spacing-y-4 sm:w-full">
+            <thead className="font-black text-lg">
+              <tr>
+                <th className="pr-40 sm:pr-0">날짜</th>
+                <th>충전 금액</th>
+              </tr>
+            </thead>
+            <tbody className="font-thin text-base">
+              <tr>
+                <td></td>
+                <td></td>
+              </tr>
+            </tbody>
+          </table>
         )}
       </BackgroundCover>
       {/* absolute를 이용하여 자유롭게 위치를 조정 할 수 있도록 제작 */}
@@ -265,26 +153,6 @@ export const MyPage = () => {
     </div>
   );
 };
-
-const EditPosition = styled.label`
-  position: absolute;
-  top: 50%;
-  right: 3%;
-  transform: translateY(-50%);
-`;
-
-const ProfilePseudoElements = styled.label`
-  &::after {
-    content: "";
-    width: 1px;
-    height: 20px;
-    position: absolute;
-    right: 50%;
-    top: 50%;
-    transform: translateY(-50%);
-    background-color: var(--secondary);
-  }
-`;
 
 const ModalPosition = styled.div`
   position: absolute;
@@ -306,43 +174,7 @@ const ModalPosition = styled.div`
   }
 `;
 
-const ChargeList = styled.div`
-  width: 50%;
-  &::after {
-    content: "";
-    width: 1px;
-    height: 100%;
-    position: absolute;
-    left: 46%;
-    top: 50%;
-    background-color: var(--secondary);
-    transform: translateY(-50%);
-  }
-
-  @media (min-width: 360px) {
-    width: 100%;
-    justify-content: center;
-
-    &::after {
-      left: 40%;
-    }
-  }
-
-  @media (min-width: 480px) {
-    &::after {
-      left: 36%;
-    }
-  }
-
-  @media (min-width: 768px) {
-    width: 50%;
-    &::after {
-      left: 41%;
-    }
-  }
-`;
-
-const ShowMyOfficeButotn = styled.div`
+const ShowMyOfficeButton = styled.div`
   position: absolute;
   top: -40%;
   right: 1%;
