+++ src/pages/customer/AddOfficeReviews.tsx
@@ -40,7 +40,7 @@ export const AddOfficeReviews = () => {
       {officeData ? (
         <OfficeProfile officeData={officeData} />
       ) : (
-        <WrongApproachAlertCompo textString={`잘못된 접근입니다`} />
+        <WrongApproachAlertCompo textString={`잘못된 접근입니다`} top="80%" />
       )}
       {officeData ? (
         <form
